% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aulas.R
\name{Baixarapostilaoulivro}
\alias{Baixarapostilaoulivro}
\title{Baixar a apostila do GAE/UNIRIO}
\usage{
Baixarapostilaoulivro(apostila)
}
\arguments{
\item{apostila}{Baixar a apostila do GAE/UNIRIO}
}
\value{
Donwload da Apostila do Prof. Felipe Rafael
}
\description{
Baixar a apostila do GAE/UNIRIO
}
