% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aulas.R
\name{instalarpacote}
\alias{instalarpacote}
\title{instalar todos os pacotes do RCommander (Rcmdr), do Radiant e do Deducer}
\usage{
instalarpacote(Pacote)
}
\arguments{
\item{Pacote}{instalar todos os pacotes do RCommander (Rcmdr), do Radiant e do Deducer}
}
\value{
pacote instalado
}
\description{
instalar todos os pacotes do RCommander (Rcmdr), do Radiant e do Deducer
}
