% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{site}
\alias{site}
\title{visita ao site do GAE ou ao site do DATAUNIRIO}
\usage{
site(SITE)
}
\arguments{
\item{SITE}{visita ao site do GAE ou ao site do DATAUNIRIO}
}
\value{
abre o navegador e entra no site escolhido
}
\description{
visita ao site do GAE ou ao site do DATAUNIRIO
}
