% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aulas.R
\name{bd}
\alias{bd}
\title{Carrega o banco de dados em portugues}
\usage{
bd(CARROSEPAISES)
}
\arguments{
\item{CARROSEPAISES}{Carrega o banco de dados em portugues. Todavia voc?? precisa escrever data(mtcars) ou data(iris) antes para funcionar e remove(mtcars) para remover o original}
}
\value{
Banco de dados chamados CARROS, PAISES ou WBDATA (Banco de dados do WorldBank - precisa de Internet)
}
\description{
Carrega o banco de dados em portugues
}
